fix: remove leftover cleanup artifact
