+++ backend/routes/checkin.js
@@ -2,7 +2,7 @@ const express = require('express');
 const multer = require('multer');
 const { body, validationResult } = require('express-validator');
 const router = express.Router();
-const databaseService = require('../services/databaseService');
+const reservationService = require('../services/reservationService');
 const emailService = require('../services/emailService');
 const { supabaseAdmin } = require('../config/supabase');
 
@@ -32,7 +32,7 @@ router.get('/:token', async (req, res) => {
     }
 
     // Get comprehensive reservation details using guest dashboard data method
-    const dashboardData = await databaseService.getGuestDashboardData(token);
+    const dashboardData = await reservationService.getGuestDashboardData(token);
     
     if (!dashboardData) {
       return res.status(404).json({ error: 'Invalid or expired check-in link' });
@@ -149,14 +149,14 @@ router.post('/:token/submit',
       }
 
       // Get reservation from database
-      const reservation = await databaseService.getReservationByToken(token);
+      const reservation = await reservationService.getReservationByToken(token);
       
       if (!reservation) {
         return res.status(404).json({ error: 'Invalid or expired check-in link' });
       }
 
       // Check if check-in is already completed
-      const existingCheckin = await databaseService.getGuestCheckinByReservationId(reservation.id);
+      const existingCheckin = await reservationService.getGuestCheckinByReservationId(reservation.id);
       
       // Allow modification if isModification flag is set
       const isModification = req.body.isModification === true || req.body.isModification === 'true';
@@ -182,7 +182,7 @@ router.post('/:token/submit',
       };
 
       // Update reservation with guest information
-      const updatedReservation = await databaseService.updateReservationGuestInfo(reservation.id, guestInfoForDb);
+      const updatedReservation = await reservationService.updateReservationGuestInfo(reservation.id, guestInfoForDb);
 
       // Send confirmation email to guest
       try {
@@ -259,7 +259,7 @@ router.post('/:token/resend-invitation', async (req, res) => {
     const { token } = req.params;
     
     // Get reservation by token
-    const reservation = await databaseService.getReservationByToken(token);
+    const reservation = await reservationService.getReservationByToken(token);
     
     if (!reservation) {
       return res.status(404).json({ error: 'Invalid or expired check-in link' });
@@ -287,14 +287,14 @@ router.get('/:token/status', async (req, res) => {
     const { token } = req.params;
     
     // Get reservation by token
-    const reservation = await databaseService.getReservationByToken(token);
+    const reservation = await reservationService.getReservationByToken(token);
     
     if (!reservation) {
       return res.status(404).json({ error: 'Invalid or expired check-in link' });
     }
 
     // Get check-in details if exists
-    const checkin = await databaseService.getGuestCheckinByReservationId(reservation.id);
+    const checkin = await reservationService.getGuestCheckinByReservationId(reservation.id);
 
     res.status(200).json({
       reservationStatus: reservation.status,
