+++ backend/routes/webhooks.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const beds24Service = require('../services/beds24Service');
-const databaseService = require('../services/databaseService');
+const reservationService = require('../services/reservationService');
 const emailService = require('../services/emailService');
 
 // Beds24 webhook endpoint
@@ -23,14 +23,14 @@ router.post('/beds24', async (req, res) => {
     const eventId = webhookData.eventId || webhookData.id || `${Date.now()}-${Math.random()}`;
     
     // Check if we've already processed this event
-    const eventExists = await databaseService.webhookEventExists(eventId);
+    const eventExists = await reservationService.webhookEventExists(eventId);
     if (eventExists) {
       console.log('Webhook event already processed:', eventId);
       return res.status(200).json({ message: 'Event already processed' });
     }
 
     // Log the webhook event to reservation_webhook_logs table
-    const loggedEvent = await databaseService.logReservationWebhook(
+    const loggedEvent = await reservationService.logReservationWebhook(
       webhookData.booking?.id || webhookData.beds24BookingId,
       webhookData,
       false
@@ -43,7 +43,7 @@ router.post('/beds24', async (req, res) => {
       // Mark event as processed (handle gracefully if column doesn't exist)
       if (loggedEvent) {
         try {
-          await databaseService.markWebhookEventProcessed(loggedEvent.id);
+          await reservationService.markWebhookEventProcessed(loggedEvent.id);
         } catch (error) {
           console.log('Note: Webhook event processed successfully, but logging update failed (non-critical):', error.message);
         }
@@ -104,7 +104,7 @@ async function handleNewBooking(webhookData) {
     });
     
     // Check if reservation already exists
-    const existingReservation = await databaseService.getReservationByBeds24Id(
+    const existingReservation = await reservationService.getReservationByBeds24Id(
       bookingInfo.beds24BookingId
     );
     
@@ -112,7 +112,7 @@ async function handleNewBooking(webhookData) {
       console.log('Reservation already exists, updating with new data:', bookingInfo.beds24BookingId);
       
       // Update existing reservation with new data (handles extensions, modifications, etc.)
-      const updatedReservation = await databaseService.updateReservation(
+      const updatedReservation = await reservationService.updateReservation(
         existingReservation.id, 
         bookingInfo
       );
@@ -133,7 +133,7 @@ async function handleNewBooking(webhookData) {
     }
 
     // Create new reservation in database with complete field mapping
-    const reservation = await databaseService.createReservation(bookingInfo);
+    const reservation = await reservationService.createReservation(bookingInfo);
     
     console.log(`Created new reservation: ${reservation.id} for Beds24 booking: ${bookingInfo.beds24BookingId}`);
     
@@ -157,7 +157,7 @@ async function handleBookingUpdate(webhookData) {
     });
     
     // Find existing reservation
-    const existingReservation = await databaseService.getReservationByBeds24Id(
+    const existingReservation = await reservationService.getReservationByBeds24Id(
       bookingInfo.beds24BookingId
     );
     
@@ -168,7 +168,7 @@ async function handleBookingUpdate(webhookData) {
     }
 
     // Update reservation with new data
-    const updatedReservation = await databaseService.updateReservation(
+    const updatedReservation = await reservationService.updateReservation(
       existingReservation.id, 
       bookingInfo
     );
@@ -192,13 +192,13 @@ async function handleBookingCancellation(webhookData) {
     });
     
     // Find existing reservation
-    const existingReservation = await databaseService.getReservationByBeds24Id(
+    const existingReservation = await reservationService.getReservationByBeds24Id(
       bookingInfo.beds24BookingId
     );
     
     if (existingReservation) {
       // Update status to cancelled
-      await databaseService.updateReservationStatus(existingReservation.id, 'cancelled');
+      await reservationService.updateReservationStatus(existingReservation.id, 'cancelled');
       console.log(`Cancelled reservation: ${existingReservation.id} for Beds24 booking: ${bookingInfo.beds24BookingId}`);
     } else {
       console.log(`Reservation not found for cancellation: ${bookingInfo.beds24BookingId}`);
