+++ backend/services/utils/dbHelpers.js
@@ -1,0 +1,10 @@
+function cleanObject(obj) {
+  Object.keys(obj).forEach((key) => {
+    if (obj[key] === undefined) {
+      delete obj[key];
+    }
+  });
+  return obj;
+}
+
+module.exports = { cleanObject };
